+++ src/atoms/PowerPlugIcon/PowerPlugIcon.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default function PowerPlugIcon() {
+	return (
+		<svg xmlns="http://www.w3.org/2000/svg" width="22" height="19.6" viewBox="0 0 22 19.6">
+			<path fill="currentColor" d="M22,12.2c0-0.7-0.6-1.2-1.2-1.2h-2.4v2.4h2.4C21.4,13.4,22,12.9,22,12.2z"/>
+			<path fill="currentColor" d="M20.8,15.9h-2.4v2.4h2.4c0.7,0,1.2-0.6,1.2-1.2C22,16.4,21.4,15.9,20.8,15.9z"/>
+			<path fill="currentColor" d="M11,12.2H8.6v4.9H11c0,1.3,1.1,2.4,2.4,2.4h3.7V9.8h-3.7C12.1,9.8,11,10.9,11,12.2z"/>
+			<path fill="currentColor" d="M2.4,11c0-1.3,1.1-2.4,2.4-2.4h1.8c2.4,0,4.3-1.9,4.3-4.3S9.1,0,6.7,0H2.4C1.8,0,1.2,0.5,1.2,1.2s0.5,1.2,1.2,1.2h4.3c1,0,1.8,0.8,1.8,1.8S7.7,6.1,6.7,6.1H4.9C2.2,6.1,0,8.3,0,11s2.2,4.9,4.9,4.9h2.4v-2.4H4.9C3.5,13.4,2.4,12.3,2.4,11z"/>
+		</svg>
+	);
+}