+++ src/pages/Settings/Settings.js
@@ -5,6 +5,7 @@ import SettingsTemplate from '../../templates/Settings/Settings';
 import PageNav from '../../molecules/PageNav/PageNav';
 
 import FireIcon from '../../atoms/FireIcon/FireIcon';
+import PowerPlugIcon from '../../atoms/PowerPlugIcon/PowerPlugIcon';
 import WaterIcon from '../../atoms/WaterIcon/WaterIcon';
 import CogwheelIcon from '../../atoms/CogwheelIcon/CogwheelIcon';
 
@@ -17,6 +18,12 @@ export default function Settings() {
 					to: '/minder-gas-nl',
 					label: 'MinderGas.nl',
 					description: 'Monitor your natural gas consumption, by connecting to mindergas.nl'
+				},
+				{
+					icon: PowerPlugIcon,
+					to: '/smart-meter',
+					label: 'Smart Meter',
+					description: 'Show information about your smart meter and gas meter'
 				},
 				{
 					icon: WaterIcon,